Sample Code from Today's Demos
Program.cs includes the code snippets we looked at and created. The earlier demo was moved to the FileIO method while the current Main executes our theater json manipulation. (Sorry for the janky formatting, I'll be updating my settings to standardize spacing and alignment in the future.)

JsonClasses.cs contains class definitions for the json objects we were working with.

Demo.json includes the data.json from the lab

Theater.json includes what we created in class

A .drawio file is included that can be opened at app.diagrams.net for reference and further updates on your own